+++ src/Bootstrap.php
@@ -43,7 +43,7 @@ class Bootstrap {
 	public function __construct( $file ) {
 		$this->file    = $file;
 		$this->dir     = dirname( $file );
-		self::$options = get_site_option( 'wp_debugging', [] );
+		self::$options = get_site_option( 'wp_debugging', [ 'wp_debug' => '1' ] );
 		@ini_set( 'display_errors', 1 );
 	}
 
@@ -74,7 +74,7 @@ class Bootstrap {
 		add_filter(
 			'wp_dependency_timeout',
 			function ( $timeout, $source ) {
-				$timeout = basename( $this->dir ) !== $source ? $timeout : 30;
+				$timeout = basename( $this->dir ) !== $source ? $timeout : 45;
 
 				return $timeout;
 			},
@@ -114,7 +114,7 @@ class Bootstrap {
 	 */
 	public function deactivate() {
 		$config_transformer = new \WPConfigTransformer( ABSPATH . 'wp-config.php' );
-		$constants          = [ 'wp_debug_log', 'script_debug', 'savequeries', 'wp_debug', 'wp_debug_display' ];
+		$constants          = [ 'wp_debug_log', 'script_debug', 'savequeries', 'wp_debug', 'wp_debug_display', 'wp_disable_fatal_error_handler' ];
 		foreach ( $constants as $constant ) {
 			$config_transformer->remove( 'constant', strtoupper( $constant ) );
 		}