+++ src/Settings.php
@@ -232,6 +232,18 @@ class Settings {
 				'title' => esc_html__( 'Set WP_DEBUG_DISPLAY to false, default is true.', 'wp-debugging' ),
 			]
 		);
+
+		add_settings_field(
+			'wp_disable_fatal_error_handler',
+			null,
+			[ $this, 'checkbox_setting' ],
+			'wp_debugging',
+			'wp_debugging',
+			[
+				'id'    => 'wp_disable_fatal_error_handler',
+				'title' => esc_html__( 'Set WP_DISABLE_FATAL_ERROR_HANDLER to true.', 'wp-debugging' ),
+			]
+		);
 	}
 
 	/**
@@ -273,7 +285,7 @@ class Settings {
 		<div class="wrap">
 			<h1><?php esc_html_e( 'WP Debugging', 'wp-debugging' ); ?></h1>
 			<div class="updated fade">
-				<p><?php echo wp_kses_post( __( '<strong>Please note:</strong> Your <code>wp-config.php</code> file must be writable by the filesystem. If your <code>wp-config.php</code> is not writable then nothing will happen. Debug constants per <a href="https://codex.wordpress.org/Debugging_in_WordPress">Debugging in WordPress</a>.', 'wp-debugging' ) ); ?></p>
+				<p><?php echo wp_kses_post( __( '<strong>Please note:</strong> Your <code>wp-config.php</code> file must be writable by the filesystem. Any errors will result in a PHP Exception being thrown. Debug constants per <a href="https://codex.wordpress.org/Debugging_in_WordPress">Debugging in WordPress</a>.', 'wp-debugging' ) ); ?></p>
 			</div>
 			<div>
 			<form method="post" action="<?php esc_attr_e( $action ); ?>">